+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/metrics/record/MetricRecord.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.aklivity.zilla.runtime.engine.internal.metrics;
+package io.aklivity.zilla.runtime.engine.metrics.record;
 
 public interface MetricRecord
 {
